extract tropo layers using dask
