+++ tools/ARIAtools.egg-info/PKG-INFO
@@ -1,5 +1,5 @@
 Metadata-Version: 2.1
 Name: ARIAtools
-Version: 1.1.5
+Version: 1.1.6
 Summary: This is the ARIA tools package without RelaxIV support
 License-File: LICENSE
