+++ tools/ARIAtools/extractProduct.py
@@ -420,7 +420,7 @@ def prep_dem(demfilename, bbox_file, prods_TOTbbox, prods_TOTbbox_metadatalyr,
         update_file.SetProjection(proj)
         del update_file
         ds_aria = gdal.Translate(f'{aria_dem}.vrt', aria_dem, format='VRT')
-        log.info('Applied cutline to produce 3 arc-sec SRTM DEM: %s', aria_dem)
+        log.info('Applied cutline to produce 3 arc-sec Copernicus GLO90 DEM: %s', aria_dem)
 
     # Load DEM and setup lat and lon arrays
     # pass expanded DEM for metadata field interpolation
@@ -658,6 +658,7 @@ def prep_metadatalayers(outname, metadata_arr, dem, key, layers,
     ifg = os.path.basename(outname)
     out_dir = os.path.dirname(outname)
     ref_outname = deepcopy(outname)
+
     # ionosphere layer, heights do not exist to exit
     if metadata_arr[0].split('/')[-1] == 'ionosphere':
         gdal.BuildVRT(outname + '.vrt', metadata_arr)
@@ -669,17 +670,17 @@ def prep_metadatalayers(outname, metadata_arr, dem, key, layers,
         model_name = metadata_arr[0].split('/')[-3]
         out_dir = os.path.join(out_dir, model_name)
         outname = os.path.join(out_dir, ifg)
-        if not os.path.exists(out_dir):
-            os.mkdir(out_dir)
+        os.makedirs(out_dir, exist_ok=True)
 
     # Get height values
     zdim = gdal.Open(metadata_arr[0]).GetMetadataItem(
         'NETCDF_DIM_EXTRA')[1:-1]
+
     hgt_field = f'NETCDF_DIM_{zdim}_VALUES'
     # Check if height layers are consistent
     if not os.path.exists(outname + '.vrt') and \
-        not len(set([gdal.Open(i).GetMetadataItem(hgt_field)
-                     for i in metadata_arr])) == 1:
+            not len(set([gdal.Open(i).GetMetadataItem(hgt_field)
+                        for i in metadata_arr])) == 1:
         raise Exception('Inconsistent heights for '
                         'metadata layer(s) ', metadata_arr,
                         ' corresponding heights: ',
@@ -690,8 +691,8 @@ def prep_metadatalayers(outname, metadata_arr, dem, key, layers,
     if 'tropo' in key or key == 'solidEarthTide':
         # get ref and sec paths
         date_dir = os.path.join(out_dir, 'dates')
-        if not os.path.exists(date_dir):
-            os.mkdir(date_dir)
+        os.makedirs(date_dir, exist_ok=True)
+
         ref_outname = os.path.join(date_dir, ifg.split('_')[0])
         sec_outname = os.path.join(date_dir, ifg.split('_')[1])
         ref_str = 'reference/' + key
@@ -704,13 +705,14 @@ def prep_metadatalayers(outname, metadata_arr, dem, key, layers,
 
         # write ref and sec files
         for i in tup_outputs:
+            dst = f'{i[0]}.vrt'
             # delete temporary files to circumvent potential inconsistent dims
             for j in glob.glob(i[0]+'*'):
                 os.remove(j)
-            gdal.BuildVRT(i[0]+'.vrt', i[1])
+            gdal.BuildVRT(dst, i[1])
 
             # write height layers
-            gdal.Open(i[0]+'.vrt').SetMetadataItem(hgt_field,
+            gdal.Open(dst).SetMetadataItem(hgt_field,
                                                    gdal.Open(i[1][0]).GetMetadataItem(hgt_field))
 
         # compute differential
@@ -726,11 +728,13 @@ def prep_metadatalayers(outname, metadata_arr, dem, key, layers,
                     da.rio.to_raster(i, driver=driver)
                     da.close()
     else:
-        if not os.path.exists(outname+'_merged.vrt'):
-            gdal.BuildVRT(outname+'_merged.vrt', metadata_arr)
+        # src = f'{outname}_merged.vrt'
+        src = f'{outname}.vrt'
+        if not os.path.exists(src):
+            gdal.BuildVRT(src, metadata_arr)
             # write height layers
-            gdal.Open(outname+'_merged.vrt').SetMetadataItem(hgt_field,
-                                                             gdal.Open(metadata_arr[0]).GetMetadataItem(hgt_field))
+            gdal.Open(src).SetMetadataItem(hgt_field, gdal.Open(metadata_arr[0]).GetMetadataItem(hgt_field))
+
 
     return hgt_field, model_name, ref_outname
 
@@ -788,8 +792,8 @@ def handle_epoch_layers(layers,
                         product_dict,
                         lyr_path,
                         key,
-                        sec_key,
-                        ref_key,
+                        dry_key,
+                        wet_key,
                         tropo_total,
                         workdir,
                         prog_bar,
@@ -809,24 +813,22 @@ def handle_epoch_layers(layers,
     Specifically record reference/secondary components within a `dates` subdir
     and deposit the differential fields in the level above.
     """
-
     # Depending on type, set sec/ref output dirs
     if key == 'troposphereTotal':
         layers.append(key)
-        sec_workdir = os.path.join(os.path.dirname(workdir),
-                                   sec_key)
-        ref_workdir = os.path.join(os.path.dirname(workdir),
-                                   ref_key)
+        dry_workdir = os.path.join(os.path.dirname(workdir),
+                                   dry_key)
+        wet_workdir = os.path.join(os.path.dirname(workdir),
+                                   wet_key)
     else:
-        sec_workdir = deepcopy(workdir)
-        ref_workdir = deepcopy(workdir)
+        dry_workdir = deepcopy(workdir)
+        wet_workdir = deepcopy(workdir)
     # If specified workdir doesn't exist, create it
-    all_workdirs = [workdir, sec_workdir, ref_workdir]
+    all_workdirs = [workdir, dry_workdir, wet_workdir]
     all_workdirs = list(set(all_workdirs))
     existing_outputs = []
     for i in all_workdirs:
-        if not os.path.exists(i):
-            os.mkdir(i)
+        os.makedirs(i, exist_ok=True)
         # for dedup, record previous outputs to avoid reprocessing
         existing_outputs.extend(glob.glob(os.path.join(
             i, '*/*[0-9].vrt')))
@@ -840,75 +842,80 @@ def handle_epoch_layers(layers,
 
     # Iterate through all IFGs
     all_outputs = []
-    for i in enumerate(product_dict[0]):
-        ifg = product_dict[1][i[0]][0]
-        outname = os.path.abspath(os.path.join(workdir, ifg))
+    for i, prod in enumerate(product_dict[0]):
+        ifg  = os.path.basename(prod[0].split(':')[1]).split('-')[6]
 
-        # create temp files for ref/sec components
-        ref_outname = os.path.abspath(os.path.join(ref_workdir, ifg))
-        hgt_field, model_name, ref_outname = prep_metadatalayers(ref_outname,
-                                                                 i[1], dem,
-                                                                 ref_key, layers, outputFormat)
+        # generate the differential and ref/sec date wet delay
+        ifg_outname = os.path.abspath(os.path.join(wet_workdir, ifg))
+        hgt_field, model_name, wet_outname = \
+            prep_metadatalayers(ifg_outname, prod, dem, wet_key, layers, outputFormat)
+
         # Update progress bar
-        prog_bar.update(i[0]+1, suffix=ifg)
+        if prog_bar is not None:
+            prog_bar.update(i+1, suffix=ifg)
 
         # record output directories
         if model_name is not None:
             all_outputs.append(os.path.join(workdir, model_name))
-            all_outputs.append(os.path.join(ref_workdir, model_name))
-            all_outputs.append(os.path.join(sec_workdir, model_name))
+            all_outputs.append(os.path.join(wet_workdir, model_name))
+            all_outputs.append(os.path.join(dry_workdir, model_name))
         else:
             all_outputs.append(workdir)
-            all_outputs.append(ref_workdir)
-            all_outputs.append(sec_workdir)
+            all_outputs.append(wet_workdir)
+            all_outputs.append(dry_workdir)
 
         # capture if tropo and separate distinct wet and hydro layers
         if 'tropo' in key:
-            sec_outname = os.path.abspath(os.path.join(sec_workdir, ifg))
+            dry_outname = os.path.abspath(os.path.join(dry_workdir, ifg))
+            ## just used for replacing wet with hydrostatic path
             wet_path = os.path.join(lyr_path, model_name,
-                                    'reference', ref_key)
+                                    'reference', wet_key)
             dry_path = os.path.join(lyr_path, model_name,
-                                    'reference', sec_key)
-            sec_comp = [j.replace(wet_path, dry_path)
-                        for j in i[1]]
-            hgt_field, model_name, sec_outname = prep_metadatalayers(
-                sec_outname,
-                sec_comp, dem,
-                sec_key, layers, outputFormat)
+                                    'reference', dry_key)
+            ## replace wet with hydrostatic
+            dry_comp = [j.replace(wet_path, dry_path)
+                        for j in prod]
+
+            hgt_field, model_name, dry_outname = prep_metadatalayers(
+                dry_outname,
+                dry_comp, dem,
+                dry_key, layers, outputFormat)
             # if specified, compute total delay
             if tropo_total:
                 model_dir = os.path.abspath(os.path.join(workdir, model_name))
-                outname = os.path.join(model_dir, ifg)
+                outname   = os.path.join(model_dir, ifg) # ifg name
+
                 # compute reference diff
-                ref_diff = ref_outname
-                sec_diff = sec_outname
+                wet_diff = wet_outname
+                dry_diff = dry_outname
                 outname_diff = os.path.join(model_dir, 'dates',
-                                            os.path.basename(ref_diff))
+                                            os.path.basename(wet_diff))
+
                 if not os.path.exists(outname_diff):
-                    generate_diff(ref_diff, sec_diff, outname_diff, key,
-                                  sec_key, tropo_total, hgt_field, outputFormat)
+                    generate_diff(wet_diff, dry_diff, outname_diff, key,
+                                  dry_key, tropo_total, hgt_field, outputFormat)
                 # compute secondary diff
-                ref_diff = os.path.join(os.path.dirname(ref_outname),
+                wet_diff_s = os.path.join(os.path.dirname(wet_outname),
                                         ifg.split('_')[1])
-                sec_diff = os.path.join(os.path.dirname(sec_outname),
+                dry_diff_s = os.path.join(os.path.dirname(dry_outname),
                                         ifg.split('_')[1])
-                outname_diff = os.path.join(model_dir, 'dates',
-                                            os.path.basename(ref_diff))
-                if not os.path.exists(outname_diff):
-                    generate_diff(ref_diff, sec_diff, outname_diff, key,
-                                  sec_key, tropo_total, hgt_field, outputFormat)
-                # compute total diff
-                ref_diff = os.path.join(ref_workdir, model_name, ifg)
-                sec_diff = os.path.join(sec_workdir, model_name, ifg)
-                generate_diff(ref_diff, sec_diff, outname, key,
-                              sec_key, tropo_total, hgt_field, outputFormat)
+                outname_diff_s = os.path.join(model_dir, 'dates',
+                                            os.path.basename(wet_diff_s))
+                if not os.path.exists(outname_diff_s):
+                    generate_diff(wet_diff_s, dry_diff_s, outname_diff_s, key,
+                                  dry_key, tropo_total, hgt_field, outputFormat)
+
+                # compute differential
+                wet_diff_d = os.path.join(wet_workdir, model_name, ifg)
+                dry_diff_d = os.path.join(dry_workdir, model_name, ifg)
+                generate_diff(wet_diff_d, dry_diff_d, outname, key,
+                              dry_key, tropo_total, hgt_field, outputFormat)
         else:
-            sec_outname = os.path.dirname(ref_outname)
-            sec_outname = os.path.abspath(os.path.join(sec_outname,
+            dry_outname = os.path.abspath(os.path.join(os.path.dirname(wet_outname),
                                                        ifg.split('_')[0]))
 
     # delete temporary files if layers not requested
-    prod_ver_list = i[1]
+    prod_ver_list = prod
     for i in all_workdirs:
         key_name = os.path.basename(i)
         if os.path.exists(i):
@@ -916,26 +923,30 @@ def handle_epoch_layers(layers,
                 shutil.rmtree(i)
 
     # interpolate and intersect epochs for user requested layers
-    all_outputs = list(set(all_outputs))
-    for i in enumerate(all_outputs):
-        if os.path.exists(i[1]):
+    all_outputs = sorted(list(set(all_outputs)))
+
+    for i, path_res in enumerate(all_outputs):
+        if os.path.exists(path_res):
             # Update progress bar
-            prog_bar.update(i[0]+1, suffix=ifg)
+            if prog_bar is not None:
+                prog_bar.update(i+1, suffix=ifg)
 
             record_epochs = []
-            record_epochs.extend(glob.glob(os.path.join(
-                i[1], '*[0-9].vrt')))
-            record_epochs.extend(glob.glob(os.path.join(
-                i[1], 'dates/*[0-9].vrt')))
+            record_epochs.extend(glob.glob(os.path.join(path_res, '*[0-9].vrt')))
+            record_epochs.extend(glob.glob(os.path.join(path_res, 'dates/*[0-9].vrt')))
+
             # dedup check for interpolating only new files
             record_epochs = [j for j in record_epochs
                              if j not in existing_outputs]
+
             for j in enumerate(record_epochs):
                 # Interpolate/intersect with DEM before cropping
                 finalize_metadata(j[1][:-4], bounds, dem_bounds,
                                   prods_TOTbbox, dem, lat, lon, hgt_field,
                                   prod_ver_list, mask, outputFormat,
                                   verbose=verbose)
+                # BZ continue
+                # continue
                 # If necessary, resample raster
                 if multilooking is not None:
                     resampleRaster(j[1][:-4], multilooking, bounds,
@@ -954,7 +965,8 @@ def handle_epoch_layers(layers,
                     dim_check(ref_arr, prod_arr)
                 prev_outname = j[1][:-4]
 
-    prog_bar.close()
+    if prog_bar is not None:
+        prog_bar.close()
 
     return
 
@@ -1035,14 +1047,14 @@ def export_products(full_product_dict, bbox_file, prods_TOTbbox, layers,
                                                  [layers, tropo_lyrs]))) != []:
         # set input keys
         lyr_prefix = '/science/grids/corrections/external/troposphere/'
-        key = 'troposphereTotal'
+        key     = 'troposphereTotal'
         wet_key = 'troposphereWet'
         dry_key = 'troposphereHydrostatic'
         workdir = os.path.join(outDir, key)
         lyr_input_dict['lyr_path'] = lyr_prefix
         lyr_input_dict['key'] = key
-        lyr_input_dict['sec_key'] = dry_key
-        lyr_input_dict['ref_key'] = wet_key
+        lyr_input_dict['dry_key'] = dry_key
+        lyr_input_dict['wet_key'] = wet_key
         lyr_input_dict['tropo_total'] = tropo_total
         lyr_input_dict['workdir'] = workdir
 
@@ -1082,13 +1094,14 @@ def export_products(full_product_dict, bbox_file, prods_TOTbbox, layers,
 
     # If specified, extract solid earth tides
     tropo_lyrs = list(set(tropo_lyrs))
+    tropo_lyrs.append('troposphereTotal') if tropo_total else ''
     ext_corr_lyrs = tropo_lyrs + ['solidEarthTide', 'troposphereTotal']
     if list(set.intersection(*map(set,
                                   [layers, ['solidEarthTide']]))) != []:
         lyr_prefix = '/science/grids/corrections/external/tides/solidEarth/'
         key = 'solidEarthTide'
-        ref_key = key
-        sec_key = key
+        dry_key = key
+        wet_key = key
         product_dict = \
             [[j[key] for j in full_product_dict if key in j.keys()],
              [j["pair_name"] for j in full_product_dict if key in j.keys()]]
@@ -1103,8 +1116,8 @@ def export_products(full_product_dict, bbox_file, prods_TOTbbox, layers,
         lyr_input_dict['product_dict'] = product_dict
         lyr_input_dict['lyr_path'] = lyr_prefix
         lyr_input_dict['key'] = key
-        lyr_input_dict['sec_key'] = sec_key
-        lyr_input_dict['ref_key'] = ref_key
+        lyr_input_dict['dry_key'] = dry_key
+        lyr_input_dict['wet_key'] = wet_key
         lyr_input_dict['tropo_total'] = False
         lyr_input_dict['workdir'] = workdir
 
@@ -1119,7 +1132,7 @@ def export_products(full_product_dict, bbox_file, prods_TOTbbox, layers,
         ref_arr = [ref_wid, ref_hgt, ref_geotrans,
                    prev_outname]
 
-    # Loop through other user expected layers
+    # Loop through other (non tropo/SET) user expected layers
     layers = [i for i in layers if i not in ext_corr_lyrs]
     for key_ind, key in enumerate(layers):
         product_dict = [[j[key] for j in full_product_dict],
@@ -1130,8 +1143,8 @@ def export_products(full_product_dict, bbox_file, prods_TOTbbox, layers,
             product_dict[0]), prefix='Generating: '+key+' - ')
 
         # If specified workdir doesn't exist, create it
-        if not os.path.exists(workdir):
-            os.mkdir(workdir)
+
+        os.makedirs(workdir, exist_ok=True)
 
         # Iterate through all IFGs
         # TODO can we wrap this into funtion and run it
@@ -1143,10 +1156,11 @@ def export_products(full_product_dict, bbox_file, prods_TOTbbox, layers,
             prog_bar.update(i[0]+1, suffix=ifg)
 
             # Extract/crop metadata layers
+            prod_layers = i[1] # list of paths to  specified layer in all gunws
             if any(":/science/grids/imagingGeometry"
-                   in s for s in i[1]) or \
-                any(":/science/grids/corrections"
-                    in s for s in i[1]):
+                   in s for s in prod_layers) or \
+                any(":/science/grids/corrections/derived/ionosphere/ionosphere"
+                    in s for s in prod_layers):
                 # make VRT pointing to metadata layers in standard product
                 hgt_field, model_name, outname = prep_metadatalayers(outname,
                                                                      i[1], dem, key, layers)
@@ -1294,23 +1308,22 @@ def finalize_metadata(outname, bbox_bounds, dem_bounds, prods_TOTbbox, dem,
     # import dependencies
     from scipy.interpolate import RegularGridInterpolator
 
-    # get final shape
-    # MG: replace string with gdal instance
+    # get FINAL shape
     if isinstance(dem, str):
         dem = gdal.Open(dem)
     else:
-        dem_bounds = dem.GetGeoTransform()
+        dem = gdal.Open(dem.GetDescription())
 
-    arrres = gdal.Open(dem.GetDescription())
-    arrshape = [arrres.RasterYSize, arrres.RasterXSize]
-    ref_geotrans = arrres.GetGeoTransform()
-    arrres = [abs(ref_geotrans[1]), abs(ref_geotrans[-1])]
+    dem_bounds = dem.GetGeoTransform()
+
+    arrshape = [dem.RasterYSize, dem.RasterXSize]
+    gt       = dem.GetGeoTransform()
+    arrres   = [abs(gt[1]), abs(gt[-1])]
 
     # load layered metadata array
-    tmp_name = outname+'_merged.vrt'
-    data_array = gdal.Warp('', tmp_name,
-                           options=gdal.WarpOptions(format="MEM",
-                                                    options=['NUM_THREADS=%s' % (num_threads)]))
+    # tmp_name = outname+'_merged.vrt'
+    tmp_name = outname+'.vrt'
+    data_array = gdal.Warp('', tmp_name, format="MEM", options=['NUM_THREADS=%s' % (num_threads)])
 
     # get minimum version
     version_check = []
@@ -1340,18 +1353,19 @@ def finalize_metadata(outname, bbox_bounds, dem_bounds, prods_TOTbbox, dem,
         if len(os.listdir(plots_subdir)) == 0:
             shutil.rmtree(plots_subdir)
 
+
     # Get data nodata
     data_array_nodata = data_array.GetRasterBand(1).GetNoDataValue()
     data_fmt = data_array.ReadAsArray().dtype.name
 
     # only perform DEM intersection for rasters with valid height levels
-    nohgt_lyrs = ['ionosphere']
+    nohgt_lyrs  = ['ionosphere']
+    heightsMeta = np.array(gdal.Open(f'{outname}.vrt').GetMetadataItem(
+        hgt_field)[1:-1].split(','), dtype='float32')
+
     if metadatalyr_name not in nohgt_lyrs:
         tmp_name = outname+'_temp'
         # Define lat/lon/height arrays for metadata layers
-        heightsMeta = np.array(gdal.Open(outname+'_merged.vrt').GetMetadataItem(
-            hgt_field)[1:-1].split(','), dtype='float32')
-
         latitudeMeta = np.linspace(data_array.GetGeoTransform()[3],
                                    data_array.GetGeoTransform()[3] +
                                    (data_array.GetGeoTransform()[5] *
@@ -1385,14 +1399,14 @@ def finalize_metadata(outname, bbox_bounds, dem_bounds, prods_TOTbbox, dem,
         out_interpolated = interper(pnts.transpose(2, 1, 0))
 
         # Save file
-        # MG: dem_bounds are not in use, so add small fix solution when
-        # pass dem as string
         renderVRT(tmp_name, out_interpolated,
                   geotrans=dem_bounds,
                   drivername=outputFormat,
                   gdal_fmt=data_fmt,
                   proj=dem.GetProjection(),
-                  nodata=data_array_nodata)
+                  nodata=data_array_nodata, test=False)
+
+
         del out_interpolated, interper, pnts, latitudeMeta, longitudeMeta, heightsMeta, data_array_inp
 
     dem, data_array = None, None
@@ -1402,7 +1416,7 @@ def finalize_metadata(outname, bbox_bounds, dem_bounds, prods_TOTbbox, dem,
     # it must be cut to conform with these bounds.
     # Crop to track extents
 
-    gdal_warp_kwargs = {'format': outputFormat,
+    gdal_warp_kwargs = {'format': 'VRT',
                         'cutlineDSName': prods_TOTbbox,
                         'outputBounds': bbox_bounds,
                         'dstNodata': data_array_nodata,
@@ -1412,29 +1426,24 @@ def finalize_metadata(outname, bbox_bounds, dem_bounds, prods_TOTbbox, dem,
                         'multithread': True,
                         'options': [f'NUM_THREADS={num_threads}']}
 
-    gdal.Warp(tmp_name+'_temp',
-              tmp_name,
-              options=gdal.WarpOptions(**gdal_warp_kwargs))
+    gdal.Warp(tmp_name+'_temp.vrt', f'{tmp_name}.vrt', **gdal_warp_kwargs)
 
-    # Adjust shape
     gdal_warp_kwargs = {'format': outputFormat,
                         'height': arrshape[0],
                         'width': arrshape[1],
                         'options': [f'NUM_THREADS={num_threads}']}
-    gdal.Warp(outname,
-              tmp_name+'_temp',
-              options=gdal.WarpOptions(**gdal_warp_kwargs))
+    gdal.Warp(outname, tmp_name+'_temp.vrt', **gdal_warp_kwargs)
 
     # remove temp files
     for i in glob.glob(outname+'*_temp*'):
         os.remove(i)
+
     for i in glob.glob(outname+'*_merged*'):
         os.remove(i)
 
     # Update VRT
-    gdal.Translate(outname+'.vrt',
-                   outname,
-                   options=gdal.TranslateOptions(format="VRT"))
+    gdal.Translate(outname+'.vrt', outname, format="VRT")
+
 
     # Apply mask (if specified)
     if mask is not None:
@@ -1617,9 +1626,9 @@ def gacos_correction(full_product_dict, gacos_products, bbox_file,
     if len(gacos_products) > 1:
         gacos_products = os.path.join(outDir, 'merged_GACOS')
         log.info('Stitching/storing GACOS products in %s.', gacos_products)
+
         # If specified merged directory doesn't exist, create it
-        if not os.path.exists(os.path.join(outDir, 'merged_GACOS')):
-            os.mkdir(os.path.join(outDir, 'merged_GACOS'))
+        os.makedirs(os.path.join(outDir, 'merged_GACOS'), exist_ok=True)
 
         for i in tropo_date_dict:
             # only make rsc/vrt files if valid product
@@ -1794,7 +1803,7 @@ def gacos_correction(full_product_dict, gacos_products, bbox_file,
                 np.isnan(tropo_product), 0., tropo_product)
             renderVRT(outname, tropo_product,
                       geotrans=geoT, drivername=outputFormat,
-                      gdal_fmt='float32', proj=proj, nodata=0.)
+                      gdal_fmt='float32', proj=proj, nodata=0., test=True)
 
             # check if reference and secondary scenes are written to file
             if not os.path.exists(ref_outname):
@@ -1866,7 +1875,7 @@ def main(inps=None):
                                                 verbose=inps.verbose)
 
     # Perform initial layer, product, and correction sanity checks
-    inps.layers, inps.tropo_total, \
+    inps.layers, tropo_total, \
         model_names = layerCheck(standardproduct_info.products[1],
                                  inps.layers,
                                  inps.nc_version,
@@ -1947,7 +1956,7 @@ def main(inps=None):
         'verbose': inps.verbose,
         'num_threads': inps.num_threads,
         'multilooking': inps.multilooking,
-        'tropo_total': inps.tropo_total,
+        'tropo_total': tropo_total,
         'model_names': model_names
     }
