+++ tools/ARIAtools/sequential_stitching.py
@@ -16,7 +16,7 @@ from pathlib import Path
 
 
 # import util modules
-from ARIAtools.stitiching_util import (get_GUNW_array, get_GUNW_attr,
+from ARIAtools.stitching_util import (get_GUNW_array, get_GUNW_attr,
                                        frame_overlap, combine_data_to_single,
                                        write_GUNW_array, snwe_to_extent,
                                        _nan_filled_array)
