+++ tools/ARIAtools/tsSetup_dask.py
@@ -8,6 +8,7 @@
 
 import dask
 import logging
+import os.path as op
 from pathlib import Path
 from itertools import compress
 from osgeo import gdal
@@ -20,7 +21,9 @@ from ARIAtools.tsSetup import generate_stack
 from ARIAtools.mask_util import prep_mask
 from ARIAtools.shapefile_util import open_shapefile
 from ARIAtools.sequential_stitching import product_stitch_sequential
-from ARIAtools.extractProduct import merged_productbbox, prep_dem, prep_metadatalayers, finalize_metadata
+from ARIAtools.extractProduct import merged_productbbox, prep_dem, prep_metadatalayers, finalize_metadata, handle_epoch_layers
+from ARIAtools import progBar
+import logging
 
 
 gdal.UseExceptions()
@@ -212,43 +215,10 @@ def export_unwrappedPhase(product_dict,  bbox_file, prods_TOTbbox, arres,
     # Run export jobs
     vprint(f'Run number of jobs: {len(jobs)}')
     out = dask.compute(*jobs)
-    progress(out)  # need to check how to make dask progress bar with dask
+    # progress(out)  # need to check how to make dask progress bar with dask
     # close dask
     if client:
         client.close()
-
-
-def _gdal_export(inputfiles, outname, gdal_warp_kwargs, mask=None, outputFormat='ISCE'):
-    if outputFormat == 'VRT':
-        # building the virtual vrt
-        gdal.BuildVRT(outname + "_uncropped" + '.vrt', inputfiles)
-        # building the cropped vrt
-        gdal.Warp(outname+'.vrt',
-                  outname+'_uncropped.vrt',
-                  options=gdal.WarpOptions(**gdal_warp_kwargs))
-    else:
-        # building the VRT
-        gdal.BuildVRT(outname + '.vrt', inputfiles)
-        gdal.Warp(outname,
-                  outname+'.vrt',
-                  options=gdal.WarpOptions(**gdal_warp_kwargs))
-
-        # Update VRT
-        gdal.Translate(outname+'.vrt', outname,
-                       options=gdal.TranslateOptions(format="VRT"))
-
-        # Apply mask (if specified).
-        if mask is not None:
-            if isinstance(mask, str):
-                mask_file = gdal.Open(mask)
-            else:
-                # for gdal instance, from prep_mask
-                mask_file = mask
-            update_file = gdal.Open(outname, gdal.GA_Update)
-            mask_arr = mask_file.ReadAsArray() * \
-                gdal.Open(outname + '.vrt').ReadAsArray()
-            update_file.GetRasterBand(1).WriteArray(mask_arr)
-            del update_file, mask_arr
 
 
 def exportCoherenceAmplitude(product_dict, bbox_file, prods_TOTbbox, arrres,
@@ -317,32 +287,16 @@ def exportCoherenceAmplitude(product_dict, bbox_file, prods_TOTbbox, arrres,
     # Run export jobs
     vprint(f'Run number of jobs: {len(jobs)}')
     out = dask.compute(*jobs)
-    progress(out)  # need to check how to make dask progress bar with dask
+    # progress(out)  # need to check how to make dask progress bar with dask
     # close dask
     if client:
         client.close()
 
 
-def _export_metadata(inputfiles, outname, demfile, aria_dem_dict,
-                     bounds, prods_TOTbbox, mask=None, outputFormat='ISCE',
-                     n_threads=2, verbose=True):
-
-    layer = inputfiles[0].split('/')[-1]
-
-    # make VRT pointing to metadata layers in standard product
-    hgt_field = prep_metadatalayers(outname, inputfiles,
-                                    demfile, layer, layer)[0]
-
-    # Interpolate/intersect with DEM before cropping
-    finalize_metadata(outname, bounds, aria_dem_dict['dem_bounds'],
-                      prods_TOTbbox, demfile, aria_dem_dict['Latitude'],
-                      aria_dem_dict['Longitude'], hgt_field, inputfiles,
-                      mask, outputFormat, verbose=verbose, num_threads=n_threads)
-
-
 def exportImagingGeometry(product_dict, bbox_file, prods_TOTbbox, dem, Latitude, Longitude,
-                          work_dir, layer='incidenceAngle', mask=None, outputFormat='ISCE',
-                          n_threads=1, n_jobs=1, verbose=True):
+                          work_dir, layer='incidenceAngle', mask=None,
+                          outputFormat='ISCE', n_threads=1, n_jobs=1, verbose=True):
+
     # verbose printing
     def vprint(x): return print(x) if verbose == True else None
 
@@ -400,19 +354,299 @@ def exportImagingGeometry(product_dict, bbox_file, prods_TOTbbox, dem, Latitude,
         n_threads=n_threads
     )
 
-    jobs = []
+    jobs  = []
     for product, name, outFile in zipped_jobs:
         job_dict['inputfiles'] = product[layer]
         job_dict['outname'] = str(outFile)
         job = dask.delayed(_export_metadata)(**job_dict, dask_key_name=name)
         jobs.append(job)
+
     # Run export jobs
     vprint(f'Run number of jobs: {len(jobs)} with {n_jobs} workers')
     out = dask.compute(*jobs)
+    # progress(out)  # need to check how to make dask progress bar with dask
+    # close dask
+    if client:
+        client.close()
+
+
+def exportTropoSET(product_dict, bbox_file, prods_TOTbbox, dem, Latitude, Longitude, workdir,
+                   wmodel='HRRR', layer='troposphereTotal',  mask=None,
+                   n_threads=1, n_jobs=1, verbose=True, debug=False):
+    """ Export Tropo and SET corrections.
+
+    Set debug to True and use 'assert None' with %pdb on in associated notebook
+    """
+    outputFormat = 'GTiff'
+
+     # verbose printing
+    def vprint(x): return print(x) if verbose == True else None
+
+    # Check
+    available_layers = ['troposphereWet', 'troposphereHydrostatic', 'troposphereTotal', 'SET']
+    if layer not in available_layers:
+        raise ValueError(f'Selected layer: {layer} is wrong, '
+                        f'available choices: {available_layers}')
+
+
+    # initalize multiprocessing
+    if debug:
+        client = None
+
+    elif n_jobs > 1 or len(product_dict) > 1:
+        vprint(f'Running GUNW {layer} in parallel!')
+        client = Client(processes=True,
+                        threads_per_worker=1,
+                        n_workers=n_jobs, memory_limit='20GB')
+        vprint(f'Link: {client.dashboard_link}')
+    else:
+        client = None
+
+    # Check if output files exist
+    out_dir = Path(workdir) / wmodel / layer
+    outNames = [ix['pair_name'][0] for ix in product_dict]
+    outFiles = [out_dir / name for name in outNames]
+    export_list = [not (outFile.exists()) for outFile in outFiles]
+
+    # Get only the pairs we need to run
+    product_dict = compress(product_dict, export_list)
+    outNames = compress(outNames, export_list)
+
+
+    # Get bounds of bbox and dem
+    bounds = open_shapefile(bbox_file, 0, 0).bounds
+    dem_gt = dem.GetGeoTransform()
+    dem_bounds = [dem_gt[0],
+                dem_gt[3]+(dem_gt[-1]*dem.RasterYSize),
+                dem_gt[0]+(dem_gt[1]*dem.RasterXSize),
+                dem_gt[3]]
+
+    job_dict = dict(
+        mask=mask,
+        dem=dem.GetDescription(),
+        dem_bounds=dem_bounds,
+        bounds=bounds,
+        prods_TOTbbox=prods_TOTbbox,
+        outputFormat=outputFormat,
+        verbose=verbose,
+        num_threads=n_threads,
+        lon=Longitude,
+        lat=Latitude,
+        multilooking=None,
+        rankedResampling=None
+    )
+
+    jobs  = []
+    for i, (product, name) in enumerate(zip(product_dict, outNames)):
+        if 'tropo' in layer:
+            ## the function always extracts them all
+            if i == 0:
+                print ('Extracting: troposphereWet, troposphereHydrostatic, and troposphereTotal')
+
+            job_dict = {**job_dict, 'product_dict': [[product[f'{layer}_{wmodel.upper()}']]],
+                    'lyr_path':'/science/grids/corrections/external/troposphere/',
+                    'key':'troposphereTotal',
+                    'workdir': Path(workdir, 'troposphereTotal'),
+                    'layers': 'troposphereWet troposphereHydrostatic troposphereTotal'.split()
+                    }
+
+            job_dict['wet_key'] = 'troposphereWet'
+            job_dict['dry_key'] = 'troposphereHydrostatic'
+            job_dict['tropo_total'] = True
+
+            # prog_bar = progBar.progressBar(maxValue=len(inputfiles[0]),
+                                                #    prefix=f'Generating: {model} {key} - ')
+            prog_bar = None
+            job_dict['prog_bar'] = prog_bar
+
+            job = dask.delayed(handle_epoch_layers)(**job_dict, dask_key_name=name)
+            jobs.append(job)
+            # assert None
+
+        elif 'SET' in layer:
+            pass
+
+        if debug:
+            break
+
+
+    # Run export jobs
+    vprint(f'Run number of jobs: {len(jobs)} with {n_jobs} workers')
+    out = dask.compute(*jobs)
+    # progress(out)  # need to check how to make dask progress bar with dask
+    # close dask
+    if client:
+        client.close()
+
+
+def exportTropo(product_dict, bbox_file, prods_TOTbbox, dem, Latitude, Longitude, workdir,
+                   wmodel='HRRR', layer='troposphereWet',  mask=None,
+                   n_threads=1, n_jobs=1, verbose=True, debug=False):
+
+    assert layer in 'troposphereWet troposphereHydrostatic'.split(), 'Wrong layer'
+    def vprint(x): return print(x) if verbose == True else None
+
+    # Get bounds of bbox and dem
+    bounds = open_shapefile(bbox_file, 0, 0).bounds
+    dem_gt = dem.GetGeoTransform()
+    dem_bounds = [dem_gt[0],
+                dem_gt[3]+(dem_gt[-1]*dem.RasterYSize),
+                dem_gt[0]+(dem_gt[1]*dem.RasterXSize),
+                dem_gt[3]]
+
+    # initalize multiprocessing
+    if debug:
+        client = None
+
+    elif n_jobs > 1 or len(product_dict) > 1:
+        vprint(f'Running GUNW {layer} in parallel!')
+        client = Client(processes=True,
+                        threads_per_worker=1,
+                        n_workers=n_jobs, memory_limit='20GB')
+        vprint(f'Link: {client.dashboard_link}')
+    else:
+        client = None
+
+    # Check if output files exist
+    out_dir = Path(workdir) / wmodel / layer
+    outNames = [ix['pair_name'][0] for ix in product_dict]
+    outFiles = [out_dir / name for name in outNames]
+    export_list = [not (outFile.exists()) for outFile in outFiles]
+
+    # Get only the pairs we need to run
+    # product_dict = compress(product_dict, export_list)
+    # outNames     = compress(outNames, export_list)
+
+    job_dict = dict(
+        mask=mask,
+        dem=dem.GetDescription(),
+        dem_bounds=dem_bounds,
+        bounds=bounds,
+        prods_TOTbbox=prods_TOTbbox,
+        verbose=verbose,
+        num_threads=n_threads,
+        lon=Longitude,
+        lat=Latitude,
+        layer=layer,
+        wmodel=wmodel,
+        workdir=workdir
+    )
+
+    ref_jobs, sec_jobs  = [], []
+    for i, (prod_dict, name) in enumerate(zip(product_dict, outNames)):
+        job_dict['prods'] = prod_dict[f'{layer}_{wmodel.upper()}'] # all tropo ifgs
+
+        # this will just export ref/sec and intersect ref with dem
+        job_r = dask.delayed(_export_tropo)(**job_dict, dask_key_name=name)
+        ref_jobs.append(job_r)
+
+        # this will just intersect the secondary with dem
+        job_s = dask.delayed(_export_tropo)(**job_dict, dask_key_name=name)
+        sec_jobs.append(job_s)
+
+        if debug:
+            break
+
+    # Run export jobs
+    vprint(f'Run number of reference jobs: {len(ref_jobs)} with {n_jobs} workers')
+    out = dask.compute(*ref_jobs)
+
+    ## then do secondary
+    vprint(f'Run number of secondary jobs: {len(sec_jobs)} with {n_jobs} workers')
+    out = dask.compute(*sec_jobs)
+
+    ## rename the files
+    # src_dir = op.join(hyd_workdir, ifg)
+    # dst_dir = op.join(hyd_workdir)
+    # os.rename(src_dir, dst_dir)
 
     # close dask
     if client:
         client.close()
+
+
+def _export_tropo(prods, layer, wmodel, workdir, bounds, prods_TOTbbox,
+                    dem, dem_bounds, lat, lon, num_threads, mask=None, verbose=False):
+    ifg         = op.basename(prods[0].split(':')[1]).split('-')[6]
+    ref, sec    = ifg.split('_')
+    hyd_workdir = op.join(workdir, layer)
+    ## double up the ifg so that files dont get reused
+    ifg_outname = op.abspath(op.join(hyd_workdir, ifg, ifg))
+    ref_outname = op.join(hyd_workdir, ifg, wmodel, 'dates', ref)
+
+    # generate the differential and ref/sec date wet delay for this ifg
+    # this overwrites ref and sec; if sec is ref for another, there is a problem
+    # not intersected with DEM
+    if not op.exists(ref_outname):
+        hgt_field, model_name, hyd_outname = \
+            prep_metadatalayers(ifg_outname, prods, dem,
+                                layer, layer, 'GTiff')
+        print ('Finalizing reference:', hyd_outname)
+
+    ## for secondary image
+    else:
+        hgt_field   = 'NETCDF_DIM_heightsMeta_VALUES'
+        hyd_outname = op.join(hyd_workdir, ifg, wmodel, 'dates', sec)
+        print ('Finalizing secondary:', hyd_outname)
+
+    # intersect with DEM
+    finalize_metadata(hyd_outname, bounds, dem_bounds,
+                    prods_TOTbbox, dem, lat, lon, hgt_field,
+                    prods, mask, 'GTiff', verbose, num_threads)
+
+def _export_metadata(inputfiles, outname, demfile, aria_dem_dict,
+                     bounds, prods_TOTbbox, mask=None, outputFormat='ISCE',
+                     n_threads=2, verbose=True):
+
+    layer = inputfiles[0].split('/')[-1]
+
+    # make VRT pointing to metadata layers in standard product
+    hgt_field, model_name, ref_outname = prep_metadatalayers(outname, inputfiles,
+                                    demfile, layer, [layer])
+
+    if model_name:
+        outname = op.join(op.dirname(outname), model_name, op.basename(outname))
+
+    # Interpolate/intersect with DEM before cropping
+    finalize_metadata(outname, bounds, aria_dem_dict['dem_bounds'],
+                      prods_TOTbbox, demfile, aria_dem_dict['Latitude'],
+                      aria_dem_dict['Longitude'], hgt_field, inputfiles,
+                      mask, outputFormat, verbose=verbose, num_threads=n_threads)
+
+
+
+def _gdal_export(inputfiles, outname, gdal_warp_kwargs, mask=None, outputFormat='ISCE'):
+    if outputFormat == 'VRT':
+        # building the virtual vrt
+        gdal.BuildVRT(outname + "_uncropped" + '.vrt', inputfiles)
+        # building the cropped vrt
+        gdal.Warp(outname+'.vrt',
+                  outname+'_uncropped.vrt',
+                  options=gdal.WarpOptions(**gdal_warp_kwargs))
+    else:
+        # building the VRT
+        gdal.BuildVRT(outname + '.vrt', inputfiles)
+        gdal.Warp(outname,
+                  outname+'.vrt',
+                  options=gdal.WarpOptions(**gdal_warp_kwargs))
+
+        # Update VRT
+        gdal.Translate(outname+'.vrt', outname,
+                       options=gdal.TranslateOptions(format="VRT"))
+
+        # Apply mask (if specified).
+        if mask is not None:
+            if isinstance(mask, str):
+                mask_file = gdal.Open(mask)
+            else:
+                # for gdal instance, from prep_mask
+                mask_file = mask
+            update_file = gdal.Open(outname, gdal.GA_Update)
+            mask_arr = mask_file.ReadAsArray() * \
+                gdal.Open(outname + '.vrt').ReadAsArray()
+            update_file.GetRasterBand(1).WriteArray(mask_arr)
+            del update_file, mask_arr
+
 
 
 def main(inps=None):
